product/views.py
from django.shortcuts import render
from rest_framework import serializers, viewsets

from .serializers import ProductSerializer
from .models import Product

class ProductViewSet(viewsets.ModelViewSet):
    serializer_class = ProductSerializer
    queryset = Product.objects.all()

    